Refactoring Form into its own component.
