+++ .gitignore
@@ -21,3 +21,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# this file is generated on build
+src/lib/azure-resource-types.json
