+++ src/lib/azureResourceTypes.js
