+++ src/lib/environments.js
@@ -1,0 +1,8 @@
+const environments = [
+  { abbr: "dev", name: "Development" },
+  { abbr: "test", name: "Test" },
+  { abbr: "stage", name: "Staging" },
+  { abbr: "prod", name: "Production" },
+];
+
+export default environments;
